requirements.txt
requests==2.25.1
click==7.1.2
google-cloud-bigquery[bqstorage,pandas]~=2.11.0
aiohttp==3.8.3
aiolimiter==1.0.b1
aiohttp-retry==2.4.5
google-cloud-secret-manager==2.12.6